+++ control_pkg/__init__.py
@@ -1,0 +1,3 @@
+import manipulator
+import gripper
+import scene
